requirements.txt
pyspark>=3.4.0
mysql-connector-python>=8.3.0
snowflake-connector-python>=3.10.0
python-dotenv>=1.0.1
pandas>=2.1.0
numpy>=1.26.0
tabulate>=0.9.0
pytest>=8.0.0
flake8>=7.0.0
